fixed TokenGen
Added an actual conversion from int64 to string representation of the number. string() would convert the int64 into the ASCII representation of the value.
Also removed string variable as it shadows the built-in string type, so I moved the function call into the return statement. With this change the code is still very much readable.